+++ src/scandeval/benchmark_dataset.py
@@ -293,7 +293,8 @@ class BenchmarkDataset(ABC):
             max_sequence_length=max_seq_length,
             vocabulary_size=vocab_size,
             generative=benchmarking_generative_model,
-            few_shot=benchmarking_generative_model,
+            # TODO: This will be changed when we support finetuning of generative models
+            few_shot=True,
             validation_split=self.benchmark_config.only_validation_split,
         )